knowledge-vault/requirements.txt
python-frontmatter>=1.0.0
PyYAML>=6.0
pathlib